Fix file organization
Make file organization clearer by adding a seperate file for spectral
functions instead of dumping them into the REST file.
