+++ webmo/__init__.py
@@ -1,2 +1,3 @@
-from .webmo_rest import WebMOREST, voigt_line, gauss_line, lorentz_line
+from .webmo_rest import WebMOREST
+from .webmo_spectra import voigt_line, gauss_line, lorentz_line
 __version__ = "1.1.5"
